+++ driven/storage/conversions_user.go
@@ -60,7 +60,7 @@ func accountDevicesFromStorage(item account) []model.Device {
 }
 
 func accountDeviceFromStorage(item userDevice) model.Device {
-	return model.Device{ID: item.ID, Type: item.Type, OS: item.OS,
+	return model.Device{ID: item.ID, DeviceID: item.DeviceID, Type: item.Type, OS: item.OS,
 		DateCreated: item.DateCreated, DateUpdated: item.DateUpdated}
 }
 
@@ -209,6 +209,38 @@ func profileFromStorage(item profile) model.Profile {
 	return model.Profile{ID: item.ID, PhotoURL: item.PhotoURL, FirstName: item.FirstName, LastName: item.LastName,
 		Email: item.Email, Phone: item.Phone, BirthYear: item.BirthYear, Address: item.Address, ZipCode: item.ZipCode,
 		State: item.State, Country: item.Country, DateCreated: item.DateCreated, DateUpdated: item.DateUpdated}
+}
+
+func profilesFromStorage(items []account, sa Adapter) []model.Profile {
+	if len(items) == 0 {
+		return make([]model.Profile, 0)
+	}
+
+	//prepare accounts
+	accounts := make(map[string][]model.Account, len(items))
+	for _, account := range items {
+		appOrg, _ := sa.getCachedApplicationOrganizationByKey(account.AppOrgID)
+		rAccount := accountFromStorage(account, *appOrg)
+
+		//add account to the map
+		profileAccounts := accounts[rAccount.Profile.ID]
+		if profileAccounts == nil {
+			profileAccounts = []model.Account{}
+		}
+		profileAccounts = append(profileAccounts, rAccount)
+		accounts[rAccount.Profile.ID] = profileAccounts
+	}
+
+	//prepare profiles
+	res := make([]model.Profile, len(items))
+	for i, item := range items {
+
+		profile := profileFromStorage(item.Profile)
+		profile.Accounts = accounts[item.Profile.ID]
+
+		res[i] = profile
+	}
+	return res
 }
 
 func profileToStorage(item model.Profile) profile {